Update aard_chaosmap.xml
Added a new option that will let the user track the number of times each chaos portal has been used. To work, it assumes all chaos portals are in one bag, and that the user automatically removes the chaos portal as part of their speedwalk and places it into the bag.